[ISSUE-6] Further modularized Login page
